% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{logbooks_names}
\alias{logbooks_names}
\title{Check Icelandic or English names in Logbooks}
\usage{
logbooks_names(...)
}
\arguments{
\item{...}{One or more quoted or unquoted expression separated by commas. These should be
column names in any one of the data.frames found in package Logbooks and may be a mixture of 
English and Icelandic names.}
}
\value{
Returns a data.frame of the names listed in \code{...} for both English and Icelandic.
}
\description{
\code{logbook_names} returns both the English and Icelandic words for a name found in 
package Logbooks
}
\examples{
logbooks_names("faeri")
logbooks_names(leggpokki, deep.redfish)
logbooks_names("veman", legggluggi, blalanga)
}
