% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_function.R
\name{logbooks_translate}
\alias{logbooks_translate}
\title{Translate data.frame column names in package Logbooks to English}
\usage{
logbooks_translate(data)
}
\arguments{
\item{data}{An unqoted name of a data.frame from package Logbooks}
}
\value{
A data.frame the same as \code{data}, but with English column names
instead of Icelandic.
}
\description{
Loads a data.frame from package Logbooks while simultaneously translating the 
column names from Icelandic to English.
}
\examples{
botnv_en <- logbooks_translate(botnv)
names(botnv_en)
}
