\name{LogbooksTranslate-package}
\alias{LogbooksTranslate-package}
\alias{LogbooksTranslate}
\docType{package}
\title{
\packageTitle{LogbooksTranslate}
}
\description{
\packageDescription{LogbooksTranslate}
}
\details{
This package will translate most column names from most data sets found in the Logbooks package from Icelandic to English.
}
\author{
\packageAuthor{LogbooksTranslate} Paul Frater <pnf1@hi.is>

Maintainer: \packageMaintainer{LogbooksTranslate} Paul Frater <pnf1@hi.is>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
bottom.trawls <- translate.botnv();
names(bottom.trawls);
translation(visir, vedags, gear.type, ship.number)
}
